Move likes/comments counts from header to below tags
